Convert most instructions to rv64 compatable versions.
The remaining instructions which might need modifications are:
  Floating point (which shouldn't need to change)
  JALR
  Immediate mode shifts
  Stores and loads need a little update to support new instructions

+++ src/rars/riscv/instructions/ADDI.java
@@ -32,7 +32,7 @@ public class ADDI extends ImmediateInstruction {
         super("addi t1,t2,-100", "Addition immediate: set t1 to (t2 plus signed 12-bit immediate)", "000");
     }
 
-    public int compute(int value, int immediate) {
+    public long compute(long value, long immediate) {
         return value + immediate;
     }
 }
