+++ src/rars/riscv/instructions/ANDI.java
@@ -32,7 +32,7 @@ public class ANDI extends ImmediateInstruction {
         super("andi t1,t2,-100", "Bitwise AND immediate : Set t1 to bitwise AND of t2 and sign-extended 12-bit immediate", "111");
     }
 
-    public int compute(int value, int immediate) {
+    public long compute(long value, long immediate) {
         return value & immediate;
     }
 }
