+++ src/rars/riscv/instructions/OR.java
@@ -34,7 +34,7 @@ public class OR extends Arithmetic {
                 "0000000", "110");
     }
 
-    public int compute(int value, int value2) {
+    public long compute(long value, long value2) {
         return value | value2;
     }
 }
